Cleaned batch.py of static methods and moved it to types package, adjusted batch_gateway imports accordingly and added more inline comments.
