+++ test/unit/testBatch.py
@@ -5,7 +5,7 @@ import unittest
 sys.path.append(os.path.abspath('.'))
 
 import trolley.configuration
-import trolley.batch
+import trolley.types.batch
 from mock import MagicMock, Mock, patch
 
 import trolley.exceptions.notFoundException
@@ -71,7 +71,7 @@ class TestBatch(unittest.TestCase):
     def test_retrieve_batch(self):
         trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
         trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-        response = trolley.batch.Batch.find("B-912Q61G0BRVGC")
+        response = trolley.types.batch.Batch.find("B-912Q61G0BRVGC")
         status = {"ok":"true","batch":{"id":"B-912Q61G0BRVGC","status":"open","amount":"999.00","totalPayments":1,"currency":"USD","description":"Weekly Payouts on 2017-4-4","sentAt":"null","completedAt":"null","createdAt":"2017-05-04T19:19:38.049Z","updatedAt":"2017-05-15T16:38:21.552Z","payments":{"payments":[{"id":"P-91XQ0U0B1RW5M","recipient":{"id":"R-91XPYX3V2MM1G","referenceId":"jsmith@exafmple.com","email":"jsmith@exafmple.com","name":"John Smith","status":"archived","countryCode":"null"},"method":"paypal","methodDisplay":"PayPal","status":"pending","sourceAmount":"999.00","targetAmount":"100.10","isSupplyPayment":"false","memo":"Do something amazing!","fees":"0.00","recipientFees":"0.00","exchangeRate":"1.000000","processedAt":"null","merchantFees":"0.00","sourceCurrency":"USD","sourceCurrencyName":"US Dollar","targetCurrency":"USD","targetCurrencyName":"US Dollar","compliance":{"status":"pending","checkedAt":"null"}}],"meta":{"page":1,"pages":1,"records":1}}}}
         self.assertEqual(response, status)
 
@@ -80,13 +80,13 @@ class TestBatch(unittest.TestCase):
         with self.assertRaises(trolley.exceptions.notFoundException.NotFoundException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.find("dddd")
+            response = trolley.types.batch.Batch.find("dddd")
     @patch('paymentrails.batch.Batch.find',fake_find)
     def test_retrieve_batch_None(self):
         with self.assertRaises(trolley.exceptions.invalidFieldException.InvalidFieldException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.find(None)
+            response = trolley.types.batch.Batch.find(None)
     
     @patch('paymentrails.batch.Batch.update',fake_update)
     def test_update_batch(self):
@@ -94,7 +94,7 @@ class TestBatch(unittest.TestCase):
         trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
         body = {"update_payments": [
             {"id": "P-91XQ0U0B1RW5M", "sourceAmount": 999}]}
-        response = trolley.batch.Batch.update("B-912Q61G0BRVGC", body)
+        response = trolley.types.batch.Batch.update("B-912Q61G0BRVGC", body)
         status = {"ok": "true", "object": "updated"}
         self.assertEqual(response, status)
 
@@ -105,7 +105,7 @@ class TestBatch(unittest.TestCase):
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
             body = {"update_payments": [
                 {"id": "P-91XQ0U0B1RW5M", "sourceAmount": 999}]}
-            response = trolley.batch.Batch.update("ddddd",body)
+            response = trolley.types.batch.Batch.update("ddddd",body)
 
     @patch('paymentrails.batch.Batch.update',fake_update)
     def test_update_batch_None_Batch(self):
@@ -114,18 +114,18 @@ class TestBatch(unittest.TestCase):
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
             body = {"update_payments": [
                 {"id": "P-91XQ0U0B1RW5M", "sourceAmount": 999}]}
-            response = trolley.batch.Batch.update(None,body)
+            response = trolley.types.batch.Batch.update(None,body)
     @patch('paymentrails.batch.Batch.update',fake_update)
     def test_update_batch_None_Body(self):
         with self.assertRaises(trolley.exceptions.invalidFieldException.InvalidFieldException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.update("B-ddddd",None)
+            response = trolley.types.batch.Batch.update("B-ddddd",None)
     @patch('paymentrails.batch.Batch.delete',fake_delete)
     def test_delete_batch(self):
         trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
         trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-        response = trolley.batch.Batch.delete("B-912Q61G0BRVGC")
+        response = trolley.types.batch.Batch.delete("B-912Q61G0BRVGC")
         status = {"ok": "true", "object": "deleted"}
         self.assertEqual(response, status)
 
@@ -134,18 +134,18 @@ class TestBatch(unittest.TestCase):
         with self.assertRaises(trolley.exceptions.notFoundException.NotFoundException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.delete("ddddd")
+            response = trolley.types.batch.Batch.delete("ddddd")
     @patch('paymentrails.batch.Batch.delete',fake_delete)
     def test_delete_batch_None_Batch(self):
         with self.assertRaises(trolley.exceptions.invalidFieldException.InvalidFieldException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.delete(None)
+            response = trolley.types.batch.Batch.delete(None)
     @patch('paymentrails.batch.Batch.search',fake_search)
     def test_list_allBatchesWithQueries(self):
         trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
         trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-        response = trolley.batch.Batch.search(1, 10, "f18")
+        response = trolley.types.batch.Batch.search(1, 10, "f18")
         status = {"ok":"true","batches":[{"id":"B-91XQ40VT5HF18","status":"open","amount":"900.90","totalPayments":1,"currency":"USD","description":"Weekly Payouts on 2017-4-8","sentAt":"null","completedAt":"null","createdAt":"2017-05-08T18:30:44.905Z","updatedAt":"2017-05-12T18:39:06.125Z"}],"meta":{"page":1,"pages":1,"records":1}}
         self.assertEqual(response, status)
 
@@ -154,13 +154,13 @@ class TestBatch(unittest.TestCase):
         with self.assertRaises(trolley.exceptions.invalidFieldException.InvalidFieldException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.search(1, 10, None)
+            response = trolley.types.batch.Batch.search(1, 10, None)
             
     @patch('paymentrails.batch.Batch.summary',fake_summary)
     def test_batch_summary(self):
         trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
         trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-        response = trolley.batch.Batch.summary("B-912Q61G0BRVGC")
+        response = trolley.types.batch.Batch.summary("B-912Q61G0BRVGC")
         status = {"ok":"true","batchSummary":{"id":"B-91XQ40VT5HF18","serverTime":"2017-05-16T13:34:52.026Z","status":"open","currency":"USD","description":"Weekly Payouts on 2017-4-8","sentAt":"null","completedAt":"null","createdAt":"2017-05-08T18:30:44.905Z","processed_by":"API","updatedAt":"2017-05-12T18:39:06.125Z","quoteExpiredAt":"null","errors":[],"methods":{"paypal":{"count":1,"value":900.9,"fees":0,"recipientFees":0,"merchantFees":0,"net":900.9,"accountType":"Gateway","displayName":"PayPal"},"bank-transfer":{"count":0,"value":0,"fees":0,"recipientFees":0,"merchantFees":0,"net":0,"accountType":"PaymentRails","displayName":"Bank Transfer"}},"PaymentRailsTotal":{"count":0,"value":0,"fees":0,"recipientFees":0,"merchantFees":0,"net":0},"GatewayTotal":{"count":1,"value":900.9,"fees":0,"recipientFees":0,"merchantFees":0,"net":900.9},"total":{"count":1,"value":900.9,"fees":0,"recipientFees":0,"merchantFees":0,"net":900.9},"merchantBalances":{"GatewayTotal":0,"PaymentRailsTotal":10000},"enoughFunds":"true"}}
         self.assertEqual(response, status)
 
@@ -169,13 +169,13 @@ class TestBatch(unittest.TestCase):
         with self.assertRaises(trolley.exceptions.notFoundException.NotFoundException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.summary("dddd")
+            response = trolley.types.batch.Batch.summary("dddd")
     @patch('paymentrails.batch.Batch.summary',fake_summary)
     def test_batch_summary_None(self):
         with self.assertRaises(trolley.exceptions.invalidFieldException.InvalidFieldException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.summary(None)
+            response = trolley.types.batch.Batch.summary(None)
 
     @patch('paymentrails.batch.Batch.create',fake_create)
     @patch('paymentrails.batch.Batch.generate_quote',fake_generate_quote)
@@ -185,18 +185,18 @@ class TestBatch(unittest.TestCase):
         trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
         body = {"payments": [{"recipient": {"id": "R-91XQ0PJH39U54"},
                               "sourceAmount": "65", "memo": "", "sourceCurrency": "CAD"}]}
-        response = trolley.batch.Batch.create(body)
+        response = trolley.types.batch.Batch.create(body)
         status = {"ok": "true"}
         self.assertEqual(response,status)
 
         batchId = 'B-91XQ40VT5HF18' 
 
-        response = trolley.batch.Batch.generate_quote(batchId)
+        response = trolley.types.batch.Batch.generate_quote(batchId)
         status = {"ok":"true","batch":{"id":"B-91XQ40VT5HF18","status":"open","amount":"900.90","totalPayments":"1","currency":"USD","description":"Weekly Payouts on 2017-4-8","sentAt":"null","completedAt":"null","createdAt":"2017-05-08T18:30:44.905Z","updatedAt":"2017-05-16T13:40:08.098Z"}}
         self.assertEqual(response, status)
 
 
-        response = trolley.batch.Batch.process_batch(batchId)
+        response = trolley.types.batch.Batch.process_batch(batchId)
         status = {"ok":"true","batch":{"id":"B-91XQ40VT5HF18","status":"processing","amount":"900.90","totalPayments":1,"currency":"USD","description":"Weekly Payouts on 2017-4-8","sentAt":"2017-05-16T13:41:56.149Z","completedAt":"null","createdAt":"2017-05-08T18:30:44.905Z","updatedAt":"2017-05-16T13:41:56.150Z"}}
         self.assertEqual(response, status)
     
@@ -205,7 +205,7 @@ class TestBatch(unittest.TestCase):
         with self.assertRaises(trolley.exceptions.invalidFieldException.InvalidFieldException):
             trolley.configuration.Configuration.set_public_key(TestBatch.public_key)
             trolley.configuration.Configuration.set_private_key(TestBatch.private_key)
-            response = trolley.batch.Batch.create(None)
+            response = trolley.types.batch.Batch.create(None)
 
 
 if __name__ == '__main__':
